Update Docker config for project summary v2
