--- lambdas/project_summary/README.md
@@ -1,5 +1,0 @@
-When updating the project summary template, the following actions must be completed:
-- Clear cells P22+ through AF22+ on the work date sheets (something is corrupt here)
-- Update the `TEMPLATE` constant in `project_summary.go`
-- Update the `SUMMARY_CURBS_TABLE_MAX_ROW` constant in `project_summary.go`
-- Update the `SUMMARY_SIDEWALKS_TABLE_MAX_ROW` constant in `project_summary.go`
