+++ lambdas/project_summary/documents/square_foot.go
