+++ lambdas/project_summary/documents/utils.go
