+++ lambdas/project_summary/fixtures/data.inch_foot.json
