+++ lambdas/project_summary/go.mod
@@ -1,25 +1,22 @@
-module app.bluezoneautomation.com/lambda-project-summary
+module app.bluezoneautomation.com/lambda-project-summary/v2
 
 go 1.19
 
 require (
 	github.com/aws/aws-lambda-go v1.41.0
-	github.com/aws/aws-sdk-go v1.45.24
-	github.com/google/uuid v1.3.1
-	github.com/jackc/pgx/v5 v5.4.3
-	github.com/xuri/excelize/v2 v2.8.1-0.20231006160550-07f2c6831a85
+	github.com/aws/aws-sdk-go v1.48.11
+	github.com/google/uuid v1.4.0
+	github.com/xuri/excelize/v2 v2.8.1-0.20231130163141-a16182e004e7
 )
 
 require (
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
 	github.com/richardlehane/mscfb v1.0.4 // indirect
 	github.com/richardlehane/msoleps v1.0.3 // indirect
 	github.com/xuri/efp v0.0.0-20230802181842-ad255f2331ca // indirect
-	github.com/xuri/nfp v0.0.0-20230918160701-e5a3f5b24785 // indirect
-	golang.org/x/crypto v0.13.0 // indirect
-	golang.org/x/net v0.15.0 // indirect
+	github.com/xuri/nfp v0.0.0-20230919160717-d98342af3f05 // indirect
+	golang.org/x/crypto v0.14.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
 	golang.org/x/text v0.13.0 // indirect
 )
