--- lambdas/project_summary_v2/go.mod
@@ -1,22 +1,0 @@
-module app.bluezoneautomation.com/lambda-project-summary/v2
-
-go 1.19
-
-require (
-	github.com/aws/aws-lambda-go v1.41.0
-	github.com/aws/aws-sdk-go v1.48.11
-	github.com/google/uuid v1.4.0
-	github.com/xuri/excelize/v2 v2.8.1-0.20231130163141-a16182e004e7
-)
-
-require (
-	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
-	github.com/richardlehane/mscfb v1.0.4 // indirect
-	github.com/richardlehane/msoleps v1.0.3 // indirect
-	github.com/xuri/efp v0.0.0-20230802181842-ad255f2331ca // indirect
-	github.com/xuri/nfp v0.0.0-20230919160717-d98342af3f05 // indirect
-	golang.org/x/crypto v0.14.0 // indirect
-	golang.org/x/net v0.17.0 // indirect
-	golang.org/x/text v0.13.0 // indirect
-)
